revise: move Geography from util to collecting
